Split vulkan_graphics.cpp into focused translation units
- vulkan_graphics.cpp: constructor/destructor, Begin/End, draw operations, target management, OnResize (~450 lines)
- vulkan_graphics_pipeline.cpp: vertex descriptions, renderpass creation, shader creation, default image, pipeline cache/lookup (~318 lines)
- vulkan_graphics_frame.cpp: context lifecycle, command recording, vertex/glyph buffer submission, batch flushing (~177 lines)
- vulkan_graphics_imgui.cpp: ImGui initialization (~53 lines)
- CMakeLists.txt: register the three new source files
