+++ src/graphics/vulkan/vulkan_graphics.cpp
@@ -1,86 +1,13 @@
 #include "common.h"
-#include "canyon/platform/glfw/glfw_window.h"
 #include "canyon/graphics/vulkan/vulkan_graphics.h"
-#include "canyon/graphics/color.h"
 #include "canyon/graphics/vulkan/vulkan_command_buffer.h"
-#include "canyon/graphics/vulkan/vulkan_texture.h"
 #include "canyon/graphics/vulkan/vulkan_font.h"
+#include "canyon/graphics/vulkan/vulkan_texture.h"
 #include "canyon/graphics/vulkan/vulkan_utils.h"
 #include "canyon/graphics/stb_image_write.h"
-#include "shaders/vulkan_shaders.h"
 
 #include "backends/imgui_impl_glfw.h"
 #include "backends/imgui_impl_vulkan.h"
-
-namespace {
-    void checkVkResult(VkResult err) {
-        CHECK_VK_RESULT(err);
-    }
-}
-
-namespace {
-    VkVertexInputBindingDescription getVertexBindingDescription() {
-        VkVertexInputBindingDescription vertexBindingDesc{};
-
-        vertexBindingDesc.binding = 0;
-        vertexBindingDesc.stride = sizeof(canyon::graphics::vulkan::Graphics::Vertex);
-        vertexBindingDesc.inputRate = VK_VERTEX_INPUT_RATE_VERTEX;
-
-        return vertexBindingDesc;
-    }
-
-    std::vector<VkVertexInputAttributeDescription> getVertexAttributeDescriptions() {
-        std::vector<VkVertexInputAttributeDescription> vertexAttributeDescs(3);
-
-        vertexAttributeDescs[0].binding = 0;
-        vertexAttributeDescs[0].location = 0;
-        vertexAttributeDescs[0].format = VK_FORMAT_R32G32_SFLOAT;
-        vertexAttributeDescs[0].offset = offsetof(canyon::graphics::vulkan::Graphics::Vertex, xy);
-
-        vertexAttributeDescs[1].binding = 0;
-        vertexAttributeDescs[1].location = 1;
-        vertexAttributeDescs[1].format = VK_FORMAT_R32G32_SFLOAT;
-        vertexAttributeDescs[1].offset = offsetof(canyon::graphics::vulkan::Graphics::Vertex, uv);
-
-        vertexAttributeDescs[2].binding = 0;
-        vertexAttributeDescs[2].location = 2;
-        vertexAttributeDescs[2].format = VK_FORMAT_R32G32B32A32_SFLOAT;
-        vertexAttributeDescs[2].offset = offsetof(canyon::graphics::vulkan::Graphics::Vertex, color);
-
-        return vertexAttributeDescs;
-    }
-
-    VkVertexInputBindingDescription getFontVertexBindingDescription() {
-        VkVertexInputBindingDescription vertexBindingDesc{};
-
-        vertexBindingDesc.binding = 0;
-        vertexBindingDesc.stride = sizeof(canyon::graphics::vulkan::Graphics::FontGlyphInstance);
-        vertexBindingDesc.inputRate = VK_VERTEX_INPUT_RATE_INSTANCE;
-
-        return vertexBindingDesc;
-    }
-
-    std::vector<VkVertexInputAttributeDescription> getFontVertexAttributeDescriptions() {
-        std::vector<VkVertexInputAttributeDescription> vertexAttributeDescs(3);
-
-        vertexAttributeDescs[0].binding = 0;
-        vertexAttributeDescs[0].location = 0;
-        vertexAttributeDescs[0].format = VK_FORMAT_R32G32_SFLOAT;
-        vertexAttributeDescs[0].offset = offsetof(canyon::graphics::vulkan::Graphics::FontGlyphInstance, pos);
-
-        vertexAttributeDescs[1].binding = 0;
-        vertexAttributeDescs[1].location = 1;
-        vertexAttributeDescs[1].format = VK_FORMAT_R32_UINT;
-        vertexAttributeDescs[1].offset = offsetof(canyon::graphics::vulkan::Graphics::FontGlyphInstance, glyphIndex);
-
-        vertexAttributeDescs[2].binding = 0;
-        vertexAttributeDescs[2].location = 2;
-        vertexAttributeDescs[2].format = VK_FORMAT_R32G32B32A32_SFLOAT;
-        vertexAttributeDescs[2].offset = offsetof(canyon::graphics::vulkan::Graphics::FontGlyphInstance, color);
-
-        return vertexAttributeDescs;
-    }
-}
 
 namespace canyon::graphics::vulkan {
     Graphics::Graphics(SurfaceContext& context, VkSurfaceKHR surface, uint32_t surfaceWidth, uint32_t surfaceHeight)
@@ -115,45 +42,6 @@ namespace canyon::graphics::vulkan {
             m_defaultContext.m_vertexBufferData = nullptr;
         }
         vkDestroyPipelineCache(m_surfaceContext.GetVkDevice(), m_vkPipelineCache, nullptr);
-    }
-
-    void Graphics::InitImgui(canyon::platform::Window const& window) {
-        if (m_imguiInitialized) {
-            return;
-        }
-
-        IMGUI_CHECKVERSION();
-        ImGui::CreateContext();
-        ImGui::GetIO().ConfigFlags |= ImGuiConfigFlags_DockingEnable | ImGuiConfigFlags_ViewportsEnable;
-
-        ImGui::StyleColorsDark();
-        ImGuiStyle& style = ImGui::GetStyle();
-        if (ImGui::GetIO().ConfigFlags & ImGuiConfigFlags_ViewportsEnable) {
-            style.WindowRounding = 0;
-            style.Colors[ImGuiCol_WindowBg].w = 1.0f;
-        }
-
-        auto& glfwWindow = static_cast<canyon::platform::glfw::Window const&>(window);
-
-        ImGui_ImplGlfw_InitForVulkan(glfwWindow.GetGLFWWindow(), true);
-        ImGui_ImplVulkan_InitInfo initInfo{};
-        initInfo.Instance = m_surfaceContext.GetContext().GetInstance();
-        initInfo.PhysicalDevice = m_surfaceContext.GetVkPhysicalDevice();
-        initInfo.Device = m_surfaceContext.GetVkDevice();
-        initInfo.QueueFamily = m_surfaceContext.GetVkQueueFamily();
-        initInfo.Queue = m_surfaceContext.GetVkQueue();
-        initInfo.DescriptorPool = m_surfaceContext.GetVkDescriptorPool();
-        initInfo.RenderPass = m_renderPass->GetRenderPass();
-        initInfo.Subpass = 0;
-        initInfo.MinImageCount = m_swapchain->GetImageCount();
-        initInfo.ImageCount = m_swapchain->GetImageCount();
-        initInfo.MSAASamples = VK_SAMPLE_COUNT_1_BIT;
-        initInfo.Allocator = nullptr;
-        initInfo.CheckVkResultFn = checkVkResult;
-        ImGui_ImplVulkan_Init(&initInfo);
-        ImGui_ImplVulkan_CreateFontsTexture();
-
-        m_imguiInitialized = true;
     }
 
     void Graphics::Begin() {
@@ -552,424 +440,6 @@ namespace canyon::graphics::vulkan {
         constants.xyOffset = { -1.0f, -1.0f };
         commandBuffer.PushConstants(*m_drawingShader, VK_SHADER_STAGE_VERTEX_BIT, sizeof(PushConstants), &constants);
         commandBuffer.PushConstants(*m_fontShader, VK_SHADER_STAGE_VERTEX_BIT, sizeof(PushConstants), &constants);
-    }
-
-    VkDescriptorSet Graphics::GetDescriptorSet(Texture& image) {
-        return m_drawingShader->GetDescriptorSet(image);
-    }
-
-    VkPrimitiveTopology Graphics::ToVulkan(ETopologyType type) const {
-        switch (type) {
-        default:
-            assert(false);
-        case ETopologyType::Lines:
-            return VK_PRIMITIVE_TOPOLOGY_LINE_LIST;
-        case ETopologyType::Triangles:
-            return VK_PRIMITIVE_TOPOLOGY_TRIANGLE_LIST;
-        }
-    }
-
-    VkPipelineColorBlendAttachmentState Graphics::ToVulkan(BlendMode mode) const {
-        VkPipelineColorBlendAttachmentState currentBlend{};
-        switch (mode) {
-        default:
-        case BlendMode::Replace:
-            currentBlend.colorWriteMask = VK_COLOR_COMPONENT_R_BIT | VK_COLOR_COMPONENT_G_BIT | VK_COLOR_COMPONENT_B_BIT | VK_COLOR_COMPONENT_A_BIT;
-            currentBlend.blendEnable = VK_FALSE;
-            currentBlend.srcColorBlendFactor = VK_BLEND_FACTOR_ONE;
-            currentBlend.dstColorBlendFactor = VK_BLEND_FACTOR_ZERO;
-            currentBlend.colorBlendOp = VK_BLEND_OP_ADD;
-            currentBlend.srcAlphaBlendFactor = VK_BLEND_FACTOR_ONE;
-            currentBlend.dstAlphaBlendFactor = VK_BLEND_FACTOR_ZERO;
-            currentBlend.alphaBlendOp = VK_BLEND_OP_ADD;
-            break;
-        case BlendMode::Add:
-            currentBlend.colorWriteMask = VK_COLOR_COMPONENT_R_BIT | VK_COLOR_COMPONENT_G_BIT | VK_COLOR_COMPONENT_B_BIT | VK_COLOR_COMPONENT_A_BIT;
-            currentBlend.blendEnable = VK_TRUE;
-            currentBlend.srcColorBlendFactor = VK_BLEND_FACTOR_SRC_ALPHA;
-            currentBlend.dstColorBlendFactor = VK_BLEND_FACTOR_ONE;
-            currentBlend.colorBlendOp = VK_BLEND_OP_ADD;
-            currentBlend.srcAlphaBlendFactor = VK_BLEND_FACTOR_ZERO;
-            currentBlend.dstAlphaBlendFactor = VK_BLEND_FACTOR_ONE;
-            currentBlend.alphaBlendOp = VK_BLEND_OP_ADD;
-            break;
-        case BlendMode::Alpha:
-            currentBlend.colorWriteMask = VK_COLOR_COMPONENT_R_BIT | VK_COLOR_COMPONENT_G_BIT | VK_COLOR_COMPONENT_B_BIT | VK_COLOR_COMPONENT_A_BIT;
-            currentBlend.blendEnable = VK_TRUE;
-            currentBlend.srcColorBlendFactor = VK_BLEND_FACTOR_SRC_ALPHA;
-            currentBlend.dstColorBlendFactor = VK_BLEND_FACTOR_ONE_MINUS_SRC_ALPHA;
-            currentBlend.colorBlendOp = VK_BLEND_OP_ADD;
-            currentBlend.srcAlphaBlendFactor = VK_BLEND_FACTOR_ONE;
-            currentBlend.dstAlphaBlendFactor = VK_BLEND_FACTOR_ONE_MINUS_SRC_ALPHA;
-            currentBlend.alphaBlendOp = VK_BLEND_OP_ADD;
-            break;
-        case BlendMode::Modulate:
-            currentBlend.colorWriteMask = VK_COLOR_COMPONENT_R_BIT | VK_COLOR_COMPONENT_G_BIT | VK_COLOR_COMPONENT_B_BIT | VK_COLOR_COMPONENT_A_BIT;
-            currentBlend.blendEnable = VK_TRUE;
-            currentBlend.srcColorBlendFactor = VK_BLEND_FACTOR_ZERO;
-            currentBlend.dstColorBlendFactor = VK_BLEND_FACTOR_SRC_COLOR;
-            currentBlend.colorBlendOp = VK_BLEND_OP_ADD;
-            currentBlend.srcAlphaBlendFactor = VK_BLEND_FACTOR_ZERO;
-            currentBlend.dstAlphaBlendFactor = VK_BLEND_FACTOR_ONE;
-            currentBlend.alphaBlendOp = VK_BLEND_OP_ADD;
-            break;
-        case BlendMode::Multiply:
-            currentBlend.colorWriteMask = VK_COLOR_COMPONENT_R_BIT | VK_COLOR_COMPONENT_G_BIT | VK_COLOR_COMPONENT_B_BIT | VK_COLOR_COMPONENT_A_BIT;
-            currentBlend.blendEnable = VK_TRUE;
-            currentBlend.srcColorBlendFactor = VK_BLEND_FACTOR_DST_COLOR;
-            currentBlend.dstColorBlendFactor = VK_BLEND_FACTOR_ONE_MINUS_SRC_ALPHA;
-            currentBlend.colorBlendOp = VK_BLEND_OP_ADD;
-            currentBlend.srcAlphaBlendFactor = VK_BLEND_FACTOR_DST_ALPHA;
-            currentBlend.dstAlphaBlendFactor = VK_BLEND_FACTOR_ONE_MINUS_SRC_ALPHA;
-            currentBlend.alphaBlendOp = VK_BLEND_OP_ADD;
-            break;
-        }
-        return currentBlend;
-    }
-
-    void Graphics::CreateRenderPass() {
-        {
-            VkAttachmentDescription colorAttachment{};
-            // colorAttachment.format = VK_FORMAT_R8G8B8A8_SRGB; // TODO this might have to change?
-            colorAttachment.format = VK_FORMAT_B8G8R8A8_UNORM;
-            colorAttachment.samples = VK_SAMPLE_COUNT_1_BIT;
-            colorAttachment.loadOp = VK_ATTACHMENT_LOAD_OP_LOAD;
-            colorAttachment.storeOp = VK_ATTACHMENT_STORE_OP_STORE;
-            colorAttachment.stencilLoadOp = VK_ATTACHMENT_LOAD_OP_DONT_CARE;
-            colorAttachment.stencilStoreOp = VK_ATTACHMENT_STORE_OP_DONT_CARE;
-            colorAttachment.initialLayout = VK_IMAGE_LAYOUT_PRESENT_SRC_KHR;
-            colorAttachment.finalLayout = VK_IMAGE_LAYOUT_PRESENT_SRC_KHR;
-
-            VkAttachmentReference colorAttachmentRef{};
-            colorAttachmentRef.attachment = 0;
-            colorAttachmentRef.layout = VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL;
-
-            VkSubpassDescription subpass{};
-            subpass.pipelineBindPoint = VK_PIPELINE_BIND_POINT_GRAPHICS;
-            subpass.colorAttachmentCount = 1;
-            subpass.pColorAttachments = &colorAttachmentRef;
-
-            VkSubpassDependency dependency{};
-            dependency.srcSubpass = VK_SUBPASS_EXTERNAL;
-            dependency.dstSubpass = 0;
-            dependency.srcStageMask = VK_PIPELINE_STAGE_COLOR_ATTACHMENT_OUTPUT_BIT;
-            dependency.srcAccessMask = 0;
-            dependency.dstStageMask = VK_PIPELINE_STAGE_COLOR_ATTACHMENT_OUTPUT_BIT;
-            dependency.dstAccessMask = VK_ACCESS_COLOR_ATTACHMENT_READ_BIT | VK_ACCESS_COLOR_ATTACHMENT_WRITE_BIT;
-
-            m_renderPass = RenderPassBuilder(m_surfaceContext.GetVkDevice())
-                               .AddAttachment(colorAttachment)
-                               .AddSubpass(subpass)
-                               .AddDependency(dependency)
-                               .Build();
-        }
-        {
-            // specifically for rendering to render targets
-            VkAttachmentDescription colorAttachment{};
-            colorAttachment.format = VK_FORMAT_B8G8R8A8_UNORM;
-            colorAttachment.samples = VK_SAMPLE_COUNT_1_BIT;
-            colorAttachment.loadOp = VK_ATTACHMENT_LOAD_OP_LOAD;
-            colorAttachment.storeOp = VK_ATTACHMENT_STORE_OP_STORE;
-            colorAttachment.stencilLoadOp = VK_ATTACHMENT_LOAD_OP_DONT_CARE;
-            colorAttachment.stencilStoreOp = VK_ATTACHMENT_STORE_OP_DONT_CARE;
-            colorAttachment.initialLayout = VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL;
-            colorAttachment.finalLayout = VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL;
-
-            VkAttachmentReference colorAttachmentRef{};
-            colorAttachmentRef.attachment = 0;
-            colorAttachmentRef.layout = VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL;
-
-            VkSubpassDescription subpass{};
-            subpass.pipelineBindPoint = VK_PIPELINE_BIND_POINT_GRAPHICS;
-            subpass.colorAttachmentCount = 1;
-            subpass.pColorAttachments = &colorAttachmentRef;
-
-            VkSubpassDependency dependency{};
-            dependency.srcSubpass = VK_SUBPASS_EXTERNAL;
-            dependency.dstSubpass = 0;
-            dependency.srcStageMask = VK_PIPELINE_STAGE_COLOR_ATTACHMENT_OUTPUT_BIT;
-            dependency.srcAccessMask = 0;
-            dependency.dstStageMask = VK_PIPELINE_STAGE_COLOR_ATTACHMENT_OUTPUT_BIT;
-            dependency.dstAccessMask = VK_ACCESS_COLOR_ATTACHMENT_WRITE_BIT;
-
-            m_rtRenderPass = RenderPassBuilder(m_surfaceContext.GetVkDevice())
-                                 .AddAttachment(colorAttachment)
-                                 .AddSubpass(subpass)
-                                 .AddDependency(dependency)
-                                 .Build();
-        }
-    }
-
-    void Graphics::CreateShaders() {
-        m_drawingShader = ShaderBuilder(m_surfaceContext.GetVkDevice(), m_surfaceContext.GetVkDescriptorPool())
-                              .AddPushConstant(VK_SHADER_STAGE_VERTEX_BIT, 0, sizeof(PushConstants))
-                              .AddBinding(0, VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER, 1, VK_SHADER_STAGE_FRAGMENT_BIT)
-                              .AddStage(VK_SHADER_STAGE_VERTEX_BIT, "main", drawing_shader_vert_spv, drawing_shader_vert_spv_len)
-                              .AddStage(VK_SHADER_STAGE_FRAGMENT_BIT, "main", drawing_shader_frag_spv, drawing_shader_frag_spv_len)
-                              .Build();
-
-        m_fontShader = ShaderBuilder(m_surfaceContext.GetVkDevice(), m_surfaceContext.GetVkDescriptorPool())
-                           .AddPushConstant(VK_SHADER_STAGE_VERTEX_BIT, 0, sizeof(PushConstants))
-                           .AddBinding(0, VK_DESCRIPTOR_TYPE_STORAGE_BUFFER, 1, VK_SHADER_STAGE_VERTEX_BIT)
-                           .AddBinding(1, VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER, 1, VK_SHADER_STAGE_FRAGMENT_BIT)
-                           .AddStage(VK_SHADER_STAGE_VERTEX_BIT, "main", font_vert_spv, font_vert_spv_len)
-                           .AddStage(VK_SHADER_STAGE_FRAGMENT_BIT, "main", font_frag_spv, font_frag_spv_len)
-                           .Build();
-    }
-
-    void Graphics::CreateDefaultImage() {
-        auto stagingBuffer = std::make_unique<Buffer>(m_surfaceContext, 4, VK_BUFFER_USAGE_TRANSFER_SRC_BIT, VK_MEMORY_PROPERTY_HOST_VISIBLE_BIT | VK_MEMORY_PROPERTY_HOST_COHERENT_BIT);
-
-        unsigned char const pixel[] = { 0xFF, 0xFF, 0xFF, 0xFF };
-        void* data = stagingBuffer->Map();
-        memcpy(data, pixel, 4);
-        stagingBuffer->Unmap();
-
-        const VkFormat format = VK_FORMAT_R8G8B8A8_UNORM;
-
-        m_defaultImage = std::make_unique<Texture>(m_surfaceContext, 1, 1, format, VK_IMAGE_TILING_OPTIMAL, VK_IMAGE_USAGE_TRANSFER_DST_BIT | VK_IMAGE_USAGE_SAMPLED_BIT);
-
-        auto commandBuffer = std::make_unique<CommandBuffer>(m_surfaceContext);
-        commandBuffer->BeginRecord();
-        commandBuffer->TransitionImageLayout(*m_defaultImage, format, VK_IMAGE_LAYOUT_UNDEFINED, VK_IMAGE_LAYOUT_TRANSFER_DST_OPTIMAL);
-        commandBuffer->CopyBufferToImage(*m_defaultImage, *stagingBuffer);
-        commandBuffer->TransitionImageLayout(*m_defaultImage, format, VK_IMAGE_LAYOUT_TRANSFER_DST_OPTIMAL, VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL);
-        commandBuffer->SubmitAndWait();
-    }
-
-    RenderPass& Graphics::GetCurrentRenderPass() {
-        return IsRenderTarget() ? *m_rtRenderPass : *m_renderPass;
-    }
-
-    Pipeline& Graphics::GetCurrentPipeline(ETopologyType topology) {
-        auto context = CurrentContext();
-        auto const vkTopology = ToVulkan(topology);
-        auto const blendAttachment = ToVulkan(context->m_currentBlendMode);
-        auto const vertexInputBinding = getVertexBindingDescription();
-        auto const vertexAttributeBindings = getVertexAttributeDescriptions();
-
-        auto const builder = PipelineBuilder(m_surfaceContext.GetVkDevice())
-                                 .SetPipelineCache(m_vkPipelineCache)
-                                 .SetRenderPass(GetCurrentRenderPass())
-                                 .SetShader(m_drawingShader)
-                                 .AddVertexInputBinding(vertexInputBinding)
-                                 .AddVertexAttribute(vertexAttributeBindings[0])
-                                 .AddVertexAttribute(vertexAttributeBindings[1])
-                                 .AddVertexAttribute(vertexAttributeBindings[2])
-                                 .AddColorBlendAttachment(blendAttachment)
-                                 .AddDynamicState(VK_DYNAMIC_STATE_VIEWPORT)
-                                 .AddDynamicState(VK_DYNAMIC_STATE_SCISSOR)
-                                 .SetTopology(vkTopology);
-
-        uint32_t const pipelineHash = builder.CalculateHash();
-        auto it = m_pipelines.find(pipelineHash);
-        if (std::end(m_pipelines) == it) {
-            auto const insertResult = m_pipelines.insert(std::make_pair(pipelineHash, builder.Build()));
-            it = insertResult.first;
-        }
-
-        return *it->second;
-    }
-
-    Pipeline& Graphics::GetCurrentFontPipeline() {
-        auto context = CurrentContext();
-        auto const vkTopology = VK_PRIMITIVE_TOPOLOGY_TRIANGLE_STRIP;
-        auto const blendAttachment = ToVulkan(context->m_currentBlendMode);
-        auto const vertexInputBinding = getFontVertexBindingDescription();
-        auto const vertexAttributeBindings = getFontVertexAttributeDescriptions();
-
-        auto const builder = PipelineBuilder(m_surfaceContext.GetVkDevice())
-                                 .SetPipelineCache(m_vkPipelineCache)
-                                 .SetRenderPass(GetCurrentRenderPass())
-                                 .SetShader(m_fontShader)
-                                 .AddVertexInputBinding(vertexInputBinding)
-                                 .AddVertexAttribute(vertexAttributeBindings[0])
-                                 .AddVertexAttribute(vertexAttributeBindings[1])
-                                 .AddVertexAttribute(vertexAttributeBindings[2])
-                                 .AddColorBlendAttachment(blendAttachment)
-                                 .AddDynamicState(VK_DYNAMIC_STATE_VIEWPORT)
-                                 .AddDynamicState(VK_DYNAMIC_STATE_SCISSOR)
-                                 .SetTopology(vkTopology);
-
-        uint32_t const pipelineHash = builder.CalculateHash();
-        auto it = m_fontPipelines.find(pipelineHash);
-        if (std::end(m_fontPipelines) == it) {
-            auto const insertResult = m_fontPipelines.insert(std::make_pair(pipelineHash, builder.Build()));
-            it = insertResult.first;
-        }
-
-        return *it->second;
-    }
-
-    void Graphics::BeginContext(DrawContext* context) {
-        context->m_logicalExtent = context->m_target->GetVkExtent();
-        context->m_vertexCount = 0;
-        context->m_maxVertexCount = 1024;
-        context->m_currentPipelineId = 0;
-        context->m_pendingBatch.reset();
-
-        if (!context->m_vertexBuffer) {
-            context->m_vertexBuffer = std::make_unique<Buffer>(m_surfaceContext, 1024 * sizeof(Vertex), VK_BUFFER_USAGE_TRANSFER_DST_BIT | VK_BUFFER_USAGE_VERTEX_BUFFER_BIT, VK_MEMORY_PROPERTY_HOST_VISIBLE_BIT | VK_MEMORY_PROPERTY_HOST_COHERENT_BIT);
-            context->m_vertexBufferData = static_cast<Vertex*>(context->m_vertexBuffer->Map());
-        }
-
-        if (!context->m_fontInstanceBuffer) {
-            context->m_fontInstanceBuffer = std::make_unique<Buffer>(m_surfaceContext, 1024 * sizeof(FontGlyphInstance), VK_BUFFER_USAGE_TRANSFER_DST_BIT | VK_BUFFER_USAGE_VERTEX_BUFFER_BIT, VK_MEMORY_PROPERTY_DEVICE_LOCAL_BIT);
-            context->m_fontInstanceStagingBuffer = std::make_unique<Buffer>(m_surfaceContext, 1024 * sizeof(FontGlyphInstance), VK_BUFFER_USAGE_TRANSFER_SRC_BIT, VK_MEMORY_PROPERTY_HOST_VISIBLE_BIT);
-        }
-
-        context->m_glyphCount = 0;
-
-        m_contextStack.push(context);
-        StartCommands();
-    }
-
-    void Graphics::RestartContext() {
-        auto context = m_contextStack.top();
-        FlushCommands();
-        auto cmdFence = context->m_target->GetFence().GetVkFence();
-        vkWaitForFences(m_surfaceContext.GetVkDevice(), 1, &cmdFence, VK_TRUE, UINT64_MAX);
-        vkResetFences(m_surfaceContext.GetVkDevice(), 1, &cmdFence);
-        context->m_logicalExtent = context->m_target->GetVkExtent();
-        context->m_vertexCount = 0;
-        context->m_maxVertexCount = 1024;
-        context->m_currentPipelineId = 0;
-        context->m_pendingBatch.reset();
-        context->m_glyphCount = 0;
-        StartCommands();
-    }
-
-    void Graphics::EndContext() {
-        FlushCommands();
-        m_contextStack.pop();
-    }
-
-    void Graphics::StartCommands() {
-        auto context = m_contextStack.top();
-        auto& commandBuffer = context->m_target->GetCommandBuffer();
-        commandBuffer.Reset();
-        commandBuffer.BeginRecord();
-        commandBuffer.HostWriteToVertexBarrier(*context->m_vertexBuffer);
-        if (IsRenderTarget()) {
-            commandBuffer.TransitionImageLayout(*context->m_target->GetVkImage().m_texture, context->m_target->GetVkFormat(), VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL, VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL);
-        }
-        commandBuffer.BeginRenderPass(GetCurrentRenderPass(), *context->m_target);
-
-        VkViewport viewport;
-        viewport.x = 0;
-        viewport.y = 0;
-        viewport.width = static_cast<float>(context->m_logicalExtent.width);
-        viewport.height = static_cast<float>(context->m_logicalExtent.height);
-        viewport.minDepth = 0.0f;
-        viewport.maxDepth = 1.0f;
-        commandBuffer.SetViewport(viewport);
-
-        VkRect2D scissor;
-        scissor.offset.x = 0;
-        scissor.offset.y = 0;
-        scissor.extent = context->m_target->GetVkExtent();
-        commandBuffer.SetScissor(scissor);
-
-        PushConstants pushConstants;
-        pushConstants.xyScale = { 2.0f / static_cast<float>(context->m_logicalExtent.width), 2.0f / static_cast<float>(context->m_logicalExtent.height) };
-        pushConstants.xyOffset = { -1.0f, -1.0f };
-        commandBuffer.PushConstants(*m_drawingShader, VK_SHADER_STAGE_VERTEX_BIT, sizeof(PushConstants), &pushConstants);
-        commandBuffer.BindVertexBuffer(*context->m_vertexBuffer, 0);
-    }
-
-    void Graphics::FlushPendingBatch() {
-        auto context = CurrentContext();
-        if (!context->m_pendingBatch) {
-            return;
-        }
-        auto const& batch = *context->m_pendingBatch;
-        auto& commandBuffer = context->m_target->GetCommandBuffer();
-        commandBuffer.BindVertexBuffer(*context->m_vertexBuffer, 0);
-        commandBuffer.BindDescriptorSet(*m_drawingShader, batch.m_descriptorSet, 0);
-        commandBuffer.Draw(batch.m_vertexCount, batch.m_firstVertex);
-        context->m_pendingBatch.reset();
-    }
-
-    void Graphics::FlushCommands() {
-        FlushPendingBatch();
-        auto context = CurrentContext();
-        VkFence cmdFence = context->m_target->GetFence().GetVkFence();
-        auto& commandBuffer = context->m_target->GetCommandBuffer();
-        commandBuffer.EndRenderPass();
-        commandBuffer.EndRecord();
-
-        if (context->m_glyphCount) {
-            auto uploadCommandBuffer = std::make_unique<CommandBuffer>(m_surfaceContext);
-            uploadCommandBuffer->BeginRecord();
-
-            VkBufferMemoryBarrier barrier{};
-            barrier.sType = VK_STRUCTURE_TYPE_BUFFER_MEMORY_BARRIER;
-            barrier.srcAccessMask = VK_ACCESS_HOST_WRITE_BIT;
-            barrier.dstAccessMask = VK_ACCESS_TRANSFER_READ_BIT;
-            barrier.buffer = context->m_fontInstanceStagingBuffer->GetVKBuffer();
-            barrier.offset = 0;
-            barrier.size = context->m_fontInstanceStagingBuffer->GetSize();
-
-            vkCmdPipelineBarrier(uploadCommandBuffer->GetVkCommandBuffer(),
-                                 VK_PIPELINE_STAGE_HOST_BIT,
-                                 VK_PIPELINE_STAGE_TRANSFER_BIT,
-                                 0, 0, nullptr, 1, &barrier, 0, nullptr);
-
-            VkBufferCopy copy;
-            copy.srcOffset = 0;
-            copy.dstOffset = 0;
-            copy.size = context->m_fontInstanceStagingBuffer->GetSize();
-
-            vkCmdCopyBuffer(uploadCommandBuffer->GetVkCommandBuffer(), context->m_fontInstanceStagingBuffer->GetVKBuffer(), context->m_fontInstanceBuffer->GetVKBuffer(), 1, &copy);
-
-            barrier.srcAccessMask = VK_ACCESS_TRANSFER_WRITE_BIT;
-            barrier.dstAccessMask = VK_ACCESS_VERTEX_ATTRIBUTE_READ_BIT;
-
-            vkCmdPipelineBarrier(uploadCommandBuffer->GetVkCommandBuffer(),
-                                 VK_PIPELINE_STAGE_TRANSFER_BIT,
-                                 VK_PIPELINE_STAGE_VERTEX_INPUT_BIT,
-                                 0, 0, nullptr, 1, &barrier, 0, nullptr);
-
-            uploadCommandBuffer->SubmitAndWait();
-        }
-
-        commandBuffer.Submit(cmdFence, context->m_target->GetAvailableSemaphore(), context->m_target->GetRenderFinishedSemaphore());
-    }
-
-    void Graphics::SubmitVertices(Vertex* vertices, uint32_t vertCount, ETopologyType topology, VkDescriptorSet descriptorSet) {
-        auto context = CurrentContext();
-
-        assert(vertCount <= context->m_maxVertexCount);
-
-        const uint32_t availableVertices = context->m_maxVertexCount - context->m_vertexCount;
-        if (availableVertices < vertCount) {
-            RestartContext();
-        }
-
-        const uint32_t existingVertexOffset = context->m_vertexCount;
-        memcpy(context->m_vertexBufferData + existingVertexOffset, vertices, sizeof(Vertex) * vertCount);
-
-        auto& commandBuffer = context->m_target->GetCommandBuffer();
-
-        auto const& pipeline = GetCurrentPipeline(topology);
-        if (context->m_currentPipelineId != pipeline.m_hash) {
-            FlushPendingBatch();
-            commandBuffer.BindPipeline(pipeline);
-            context->m_currentPipelineId = pipeline.m_hash;
-        }
-
-        VkDescriptorSet resolvedDescriptorSet = (descriptorSet != VK_NULL_HANDLE)
-            ? descriptorSet
-            : m_drawingShader->GetDescriptorSet(*m_defaultImage);
-
-        if (context->m_pendingBatch && context->m_pendingBatch->m_descriptorSet == resolvedDescriptorSet) {
-            context->m_pendingBatch->m_vertexCount += vertCount;
-        } else {
-            FlushPendingBatch();
-            context->m_pendingBatch = DrawContext::PendingBatch{ existingVertexOffset, vertCount, resolvedDescriptorSet };
-        }
-
-        context->m_vertexCount += vertCount;
     }
 
     void Graphics::OnResize(VkSurfaceKHR surface, uint32_t surfaceWidth, uint32_t surfaceHeight) {
